# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: BookCon API
  version: "1.0"
  description: Self-hosted eBook library & sync API. Auth via Bearer JWT (15 min) + rotating refresh tokens. Deletions are tombstones; every syncable entity carries updated_at/deleted_at.
servers:
  - url: "{base}/api/v1"
    variables:
      base:
        default: http://localhost:8000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    BookBase:
      type: object
      required: [title, format]
      properties:
        title: { type: string }
        authors: { type: array, items: { type: string }, default: [] }
        description: { type: string, default: "" }
        language: { type: string, nullable: true }
        publisher: { type: string, nullable: true }
        published_date: { type: string, nullable: true }
        series_id: { type: string, format: uuid, nullable: true }
        series_index: { type: number, nullable: true }
    Book:
      allOf:
        - $ref: "#/components/schemas/BookBase"
        - type: object
          properties:
            id: { type: string, format: uuid }
            format: { type: string, enum: [epub, pdf, cbz, cbr] }
            status: { type: string, enum: [processing, ready, failed], default: ready }
            status_message: { type: string, nullable: true }
            cover_url: { type: string, nullable: true }
            file_size_bytes: { type: integer, nullable: true }
            page_count: { type: integer, nullable: true }
            word_count: { type: integer, nullable: true }
            tag_ids: { type: array, items: { type: string, format: uuid } }
            shelf_ids: { type: array, items: { type: string, format: uuid } }
            added_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            deleted_at: { type: string, format: date-time, nullable: true }
    Shelf:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        sort_position: { type: integer }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
    Tag:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
    Series:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
    Locator:
      description: Readium Locator object (https://readium.org/architecture/models/locators/)
      type: object
      required: [href, type]
      properties:
        href: { type: string }
        type: { type: string }
        title: { type: string }
        locations: { type: object, additionalProperties: true }
        text: { type: object, additionalProperties: true }
    Annotation:
      type: object
      required: [book_id, type, locator]
      properties:
        id: { type: string, format: uuid }
        book_id: { type: string, format: uuid }
        type: { type: string, enum: [highlight, underline, area] }
        locator: { $ref: "#/components/schemas/Locator" }
        color: { type: string, default: yellow }
        note: { type: string, default: "" }
        annotation_tags: { type: array, items: { type: string }, default: [] }
        excerpt: { type: string, default: "" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
    Bookmark:
      type: object
      required: [book_id, locator]
      properties:
        id: { type: string, format: uuid }
        book_id: { type: string, format: uuid }
        locator: { $ref: "#/components/schemas/Locator" }
        label: { type: string, default: "" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
    Position:
      type: object
      required: [book_id, locator]
      properties:
        book_id: { type: string, format: uuid }
        locator: { $ref: "#/components/schemas/Locator" }
        progress_percent: { type: number, nullable: true }
        updated_at: { type: string, format: date-time }
...
    SyncPushRequest:
      type: object
      description: Batch upsert/tombstone. Each row needs client_updated_at; server LWW.
      properties:
        annotations: { type: array, items: { $ref: "#/components/schemas/Annotation" } }
        bookmarks: { type: array, items: { $ref: "#/components/schemas/Bookmark" } }
        positions: { type: array, items: { $ref: "#/components/schemas/Position" } }
        shelves: { type: array, items: { $ref: "#/components/schemas/Shelf" } }
        tags: { type: array, items: { $ref: "#/components/schemas/Tag" } }
        series: { type: array, items: { $ref: "#/components/schemas/Series" } }
    SyncPullRequest:
      type: object
      properties:
        cursors:
          type: object
          additionalProperties: { type: string }
          description: >
            entity -> opaque watermark "<updated_at>|<last_id>" (composite so rows tied at the
            boundary timestamp are never skipped; a bare date-time is accepted for legacy clients).
            Missing = full sync.
    SyncPullResponse:
      type: object
      properties:
        cursors: { type: object, additionalProperties: { type: string, format: date-time } }
        has_more: { type: boolean }
        books: { type: array, items: { $ref: "#/components/schemas/Book" } }
        annotations: { type: array, items: { $ref: "#/components/schemas/Annotation" } }
        bookmarks: { type: array, items: { $ref: "#/components/schemas/Bookmark" } }
        positions: { type: array, items: { $ref: "#/components/schemas/Position" } }
        shelves: { type: array, items: { $ref: "#/components/schemas/Shelf" } }
        tags: { type: array, items: { $ref: "#/components/schemas/Tag" } }
        series: { type: array, items: { $ref: "#/components/schemas/Series" } }
paths:
...
  /sync/push:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SyncPushRequest" }
      responses:
        "200":
          description: Per-entity accepted ids and authoritative payloads
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: object
                    additionalProperties: { type: array, items: { type: string, format: uuid } }
                  authoritative:
                    type: object
                    additionalProperties: true
  /sync/pull:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SyncPullRequest" }
      responses:
        "200": { description: Changes since cursors, content: { application/json: { schema: { $ref: "#/components/schemas/SyncPullResponse" } } } }
...